+++ src/Shared/Infrastructure/LLMChain/EmbeddingCalculator.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace ChronicleKeeper\Shared\Infrastructure\LLMChain;
 
 use PhpLlm\LlmChain\Bridge\OpenAI\Embeddings;
-use PhpLlm\LlmChain\Model\Response\AsyncResponse;
-use PhpLlm\LlmChain\Model\Response\VectorResponse;
+use PhpLlm\LlmChain\Document\Vector;
 
+use function array_map;
 use function assert;
+use function is_array;
 
 class EmbeddingCalculator
 {
@@ -20,17 +21,33 @@ class EmbeddingCalculator
     /** @return list<float> */
     public function getSingleEmbedding(string $text): array
     {
-        $vector = $this->chainFactory->createPlatform()->request(
+        $response = $this->chainFactory->createPlatform()->request(
             model: new Embeddings(),
             input: $text,
+        )->getContent();
+        assert(is_array($response) && $response[0] instanceof Vector);
+
+        return $response[0]->getData();
+    }
+
+    /**
+     * @param list<string> $texts
+     *
+     * @return list<list<float>>
+     */
+    public function getMultipleEmbeddings(array $texts): array
+    {
+        $platform = $this->chainFactory->createPlatform();
+
+        /** @var list<Vector> $response */
+        $response = $platform->request(new Embeddings(), $texts)->getContent();
+
+        /** @var list<list<float>> $embeddings */
+        $embeddings = array_map(
+            static fn (Vector $vector): array => $vector->getData(),
+            $response,
         );
 
-        assert($vector instanceof AsyncResponse);
-        $vector = $vector->unwrap();
-
-        assert($vector instanceof VectorResponse);
-        $vector = $vector->getContent()[0];
-
-        return $vector->getData();
+        return $embeddings;
     }
 }